Write model parameters into checkpoint files
Changes in the main/phylotesting.cpp and main/phylotesting.h files were made to write the parameters of all the sequence evolution models that are tested by ModelFinder are written into the checkpoint file.
Additionally, the Newick string of the initial tree as well as the corresponding model parameters are written into the log file.